+++ src/papyrus-lang-vscode/src/ExtensionConfigProvider.ts
@@ -8,6 +8,7 @@ export interface IGameConfig {
     readonly creationKitIniFiles: string[];
     readonly installPath: string;
     readonly ignoreDebuggerVersion: boolean;
+    readonly modDirectoryPath: string;              // For supporting mod managers managing the debug plugin
 }
 
 export interface IExtensionConfig {
@@ -35,7 +36,7 @@ export class ExtensionConfigProvider implements IExtensionConfigProvider {
         return this._config;
     }
 
-    dispose() {}
+    dispose() { }
 }
 
 export const IExtensionConfigProvider = createDecorator<IExtensionConfigProvider>('extensionConfigProvider');