+++ src/papyrus-lang-vscode/src/debugger/PapyrusDebugAdapterDescriptorFactory.ts
@@ -4,7 +4,6 @@ import {
     DebugAdapterDescriptor,
     debug,
     Disposable,
-    ExtensionContext,
     window,
     commands,
     Uri,
@@ -22,8 +21,8 @@ import { ICreationKitInfoProvider } from '../CreationKitInfoProvider';
 import { IExtensionConfigProvider } from '../ExtensionConfigProvider';
 import { take } from 'rxjs/operators';
 import { IPapyrusDebugSession } from './PapyrusDebugSession';
-import { toCommandLineArgs, getGameIsRunning, debugToolPath } from '../Utilities';
-import { IExtensionContext } from '../common/vscode/IocDecorators';
+import { toCommandLineArgs, getGameIsRunning } from '../Utilities';
+import { IPathResolver } from '../common/PathResolver';
 import { IDebugSupportInstallService, DebugSupportInstallState } from './DebugSupportInstallService';
 import { ILanguageClientManager } from '../server/LanguageClientManager';
 import { showGameDisabledMessage, showGameMissingMessage } from '../features/commands/InstallDebugSupportCommand';
@@ -49,7 +48,7 @@ export class PapyrusDebugAdapterDescriptorFactory implements DebugAdapterDescrip
     private readonly _languageClientManager: ILanguageClientManager;
     private readonly _creationKitInfoProvider: ICreationKitInfoProvider;
     private readonly _configProvider: IExtensionConfigProvider;
-    private readonly _context: ExtensionContext;
+    private readonly _pathResolver: IPathResolver;
     private readonly _debugSupportInstaller: IDebugSupportInstallService;
     private readonly _registration: Disposable;
 
@@ -57,13 +56,13 @@ export class PapyrusDebugAdapterDescriptorFactory implements DebugAdapterDescrip
         @ILanguageClientManager languageClientManager: ILanguageClientManager,
         @ICreationKitInfoProvider creationKitInfoProvider: ICreationKitInfoProvider,
         @IExtensionConfigProvider configProvider: IExtensionConfigProvider,
-        @IExtensionContext context: ExtensionContext,
+        @IPathResolver pathResolver: IPathResolver,
         @IDebugSupportInstallService debugSupportInstaller: IDebugSupportInstallService
     ) {
         this._languageClientManager = languageClientManager;
         this._creationKitInfoProvider = creationKitInfoProvider;
         this._configProvider = configProvider;
-        this._context = context;
+        this._pathResolver = pathResolver;
         this._debugSupportInstaller = debugSupportInstaller;
 
         this._registration = debug.registerDebugAdapterDescriptorFactory('papyrus', this);
@@ -181,7 +180,7 @@ export class PapyrusDebugAdapterDescriptorFactory implements DebugAdapterDescrip
             clientProcessId: Number.parseInt(process.env.VSCODE_PID),
         };
 
-        const toolPath = this._context.asAbsolutePath(debugToolPath);
+        const toolPath = await this._pathResolver.getDebugToolPath();
         const commandLineArgs = toCommandLineArgs(toolArguments);
 
         const outputChannel = (await this._languageClientManager.getLanguageClientHost(session.configuration.game))